Interchain-Developer-Academy-Notes/Projects/p_gordonoh-exercise-week-2-code/x/otherworld/client/cli/query_world_params.go
package cli

import (
	"context"

	"github.com/b9lab/other-world/x/otherworld/types"
	"github.com/cosmos/cosmos-sdk/client"
	"github.com/cosmos/cosmos-sdk/client/flags"
	"github.com/spf13/cobra"
)

func CmdShowWorldParams() *cobra.Command {
	cmd := &cobra.Command{
		Use:   "show-world-params",
		Short: "shows worldParams",
		Args:  cobra.NoArgs,
		RunE: func(cmd *cobra.Command, args []string) error {
			clientCtx := client.GetClientContextFromCmd(cmd)

			queryClient := types.NewQueryClient(clientCtx)

			params := &types.QueryGetWorldParamsRequest{}

			res, err := queryClient.WorldParams(context.Background(), params)
			if err != nil {
				return err
			}

			return clientCtx.PrintProto(res)
		},
	}

	flags.AddQueryFlagsToCmd(cmd)

	return cmd
}
